Refactor code out
