+++ src/components/npo-applications/modules/Filter.js
@@ -1,0 +1,34 @@
+import React, { useState } from 'react';
+import { STATUS_FILTER_TYPE } from '../../../../utils/constants/npoVerification';
+import { Popover, Button, ListChoice } from '@kiwicom/orbit-components/';
+import { ChevronDown } from '@kiwicom/orbit-components/lib/icons';
+
+const Filter = ({ onSelectedFilter }) => {
+  const [currentFilter, setCurrentFilter] = useState(STATUS_FILTER_TYPE.ALL);
+
+  return (
+    <Popover
+      content={
+        <div>
+          {Object.values(STATUS_FILTER_TYPE).map((status) => (
+            <ListChoice
+              title={status.toUpperCase()}
+              onClick={() => {
+                setCurrentFilter(status.toUpperCase());
+                onSelectedFilter(status);
+              }}
+            />
+          ))}
+        </div>
+      }
+      noPadding
+      preferredPosition="bottom"
+    >
+      <Button size="normal" iconRight={<ChevronDown />}>
+        {currentFilter.toUpperCase()}
+      </Button>
+    </Popover>
+  );
+};
+
+export default Filter;
