+++ src/components/npo-applications/pages/NpoApplicationsPage.js
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react';
-import { Stack, Badge, Button, Popover, ListChoice } from '@kiwicom/orbit-components/';
+import { Stack, Button } from '@kiwicom/orbit-components/';
 import Table, { TableHead, TableBody, TableRow, TableCell } from '@kiwicom/orbit-components/lib/Table';
-import { STATUS_FILTER_TYPE, STATUS } from '../../../../utils/constants/npoVerification';
 import api from '../../../../api';
 import { getFormattedDateTime } from '../../../../utils/time/time';
-import { ChevronDown } from '@kiwicom/orbit-components/lib/icons';
 import { useRouter } from 'next/router';
+import BadgeStatus from '../modules/BadgeStatus';
+import Filter from '../modules/Filter';
 
 const NpoApplicationsPage = () => {
   const [filterStatus, setFilterStatus] = useState();
@@ -31,7 +31,7 @@ const NpoApplicationsPage = () => {
 
   const onReviewClicked = (npoApplicationId) => {
     router.push(`/npo-applications/${npoApplicationId}`);
-  }
+  };
 
   return (
     <Stack>
@@ -86,7 +86,9 @@ const NpoApplicationsPage = () => {
               </TableCell>
               <TableCell align="center" verticalAlign="baseline" whiteSpace="nowrap">
                 <Stack direction="row">
-                  <Button size="small" onClick={() => onReviewClicked(appSnapshot.id)}>Review</Button>
+                  <Button size="small" onClick={() => onReviewClicked(appSnapshot.id)}>
+                    Review
+                  </Button>
                 </Stack>
               </TableCell>
             </TableRow>
@@ -94,48 +96,6 @@ const NpoApplicationsPage = () => {
         </TableBody>
       </Table>
     </Stack>
-  );
-};
-
-const BadgeStatus = ({ status }) => {
-  if (status === STATUS.PENDING) {
-    return <Badge>{status.toUpperCase()}</Badge>;
-  } else if (status === STATUS.ACCEPTED) {
-    return <Badge type="successInverted">{status.toUpperCase()}</Badge>;
-  } else if (status === STATUS.RESUBMISSION) {
-    return <Badge type="warningInverted">{status.toUpperCase()}</Badge>;
-  } else if (status === STATUS.REJECTED) {
-    return <Badge type="criticalInverted">{status.toUpperCase()}</Badge>;
-  } else if (status === STATUS.REVIEWING) {
-    return <Badge type="dark">{status.toUpperCase()}</Badge>;
-  }
-};
-
-const Filter = ({ onSelectedFilter }) => {
-  const [currentFilter, setCurrentFilter] = useState(STATUS_FILTER_TYPE.ALL);
-
-  return (
-    <Popover
-      content={
-        <div>
-          {Object.values(STATUS_FILTER_TYPE).map((status) => (
-            <ListChoice
-              title={status.toUpperCase()}
-              onClick={() => {
-                setCurrentFilter(status.toUpperCase());
-                onSelectedFilter(status);
-              }}
-            />
-          ))}
-        </div>
-      }
-      noPadding
-      preferredPosition="bottom"
-    >
-      <Button size="normal" iconRight={<ChevronDown />}>
-        {currentFilter.toUpperCase()}
-      </Button>
-    </Popover>
   );
 };
